tuneApp: fix changing DOL field through CA link
Changing via Db link is not allowed for some reason,
as per AJ reply:

-----------------------------------------------------
Hi Kristi,

On Wednesday 19 January 2011 15:27:24 Luchini,
Kristi L. wrote:
>
> I can do dbpf() from the console port to change
> a DOL field, so is there a reason why dbPutLink
> can't or am I just using this incorrectly? Any
> suggestions?

You can't change a link field using a database link
(I'm not quite sure why we don't allow that but it's
probably because of the locking needed), but you can
do it with a CA link, so try adding the CA flag to
your genSub OUTx field. Note that changing links is
not a fast operation though, in some cases putting
a selRecord in front of your DOL link instead might
make more sense.

- Andrew
-----------------------------------------------------