tuneApp: fix setting a value instead of just processing Rbw
If we set the value directly we might have LINK
alarms as PINI = "YES" (in our local PV copy)
and the init routine will try to set the values
simultaneously. Instead, just force the record to
process once on initialization with the value from
autosave. If the record is already being processed
(as a result of PINI = "YES") nothing will happen.