tuneApp: fix time axis by considering the full timing scale
Previously, if we acquired less than the number
of specified acquisitions, we were assuming that
the last ones were lost and the end of the
spectrogram would be clipped by that ammount.
However, in fact, what happens is that we loose
intermediate pulses from timing, so that the end
of the spectrogram is always the same. What
changes is that the spacing between acquisitions
is larger.