tuneApp: fix redundant FreqArray-Mon record processing
Previously, we were monitoring the Trace-Mon
record (that is updated as fast as possible
according to the TuneProc IOC parameters)
to send the NORD field to NELM fiels in subArray
records. This was doen as EPICS base < 3.15 did not
send monitors for NORD fields. Since this was
fixed for EPICS base >= 3.15 we can monitor the
NORD fields directly.

This fixes #1 github issue.